day06: wip part 2 approach 1
Call me Edison, because I've found one more way not to make a lightbulb.

My approach here works perfectly for the example, but it can't forward-
project when an obstacle now will form a loop with paths that it has
not yet created. And it's pointless to have the projection turn the corner
and keep going, because it fundamentally cannot tell where the normal
path will have reached.

Another intractable approach: work backwards from the end of the map
to see if we can generate a loop in that direction, backwards. Problem:
the search space doubles literally every step of the way.

Another intractable approach: run the guard's movements, then reset
to the initial position and use this current approach to see if inserting
an obstacle generates an eventual loop. The problem here is that this
would be extremely likely to create false loops by intersecting with
a future path segment which paradoxically fails to become part of the
paths, due to the insertion of this obstacle.

I think the next approach has to be to just brute-force it, checking
whether the insertion of a new obstacle in every single point ends
up producing an infinite loop.