day09: implement part 2
I mean, I guess it makes sense for part 2 to be harder than part 1.
But this was the first day that this was really an issue!

I don't love the repeated scanning for free space, which is definitely
the cause of the massive slowdown for part 2, but I don't know any other
way to ensure that we always find the first free space that's big enough
now that we leave some free spaces open. If we were using a "real"
linked list with nodes in memory somewhere, we could keep a pointer
to the leftmost free space in the array and then scan rightwards from
there (or even: an array of 9 pointers, to the leftmost free space of size
`i-1`, which gets updated each time we put a file into one.) But
unfortunately `VecList` is pretty eager to invalidate its pointers,
and we can't rely on them sticking around like that at all.