day11: implement part 2
Ugh, they completely threw me for a loop with the emphasis on the
perfectly straight line. "Yeah, that tracks," I thought. "A `Vec`
is laid out in memory as a perfectly straight line. This is
definitely how they intended me to solve the problem." Totally
ignoring how they did not actually ask for anything which relied
on any kind of linear layout in memory.

I need to get better at thinking. When 30 seconds of thinking did
not produce the right strategy, 30 seconds of internet did... and
that was just too easy a shortcut. Not entirely proud of this one.