day16: implement the wrong answer for part 2
It works great on both examples! It _looks_ like the right answer
when I print out the solution for the real thing! But the website
claims that my answer is too high, and I don't have the patience
to figure out what's gone wrong at this point.

If nothing else, this is a reasonable starting point. It's relatively
time-efficient, for which I'm proud; I finally figured out how to
use immutable data structures for this kind of problem, after
years of trying it. They do what they say they do! They're just
not the only solution we need to prevent solve time from exploding.